backend/requirements.txt
flask
mysql-connector-python
Flask
Flask-Mail
Flask-CORS
google-generativeai
python-dotenv
